backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv==1.2.1
sqlparse==0.5.5
